* gwork: (23 commits)
  Improve readability/consistency.
  Improve readability/consistency.
  Improve constness - variable rename: don’t need Hungarian
  Use C++ style casts. Remove unnecessary casts when implicit conversion done. Few const.
  Change C casts to C++ casts for consistency.
  Fixed a couple of warnings.
  No doc timestamp. Updates all files every change.
  Style header and footer.
  Doc links.
  Added logos to docs.
  Fix documentation problems.
  Add doc link
  Fix disable for for resizer. - See #42
  Fix Window DisableResizing. - Works by disabling mouse input into resizers. - Similar attempt in GWEN: https://github.com/garrynewman/GWEN/pull/70/commits/d13fb212e8373162210f7dce7fcc6b805a1309ca   as part of https://github.com/garrynewman/GWEN/pull/70   This fix didn’t appear to work as resizers are hierarchy and not all iterated over.
  Bump latest Xcode version to 8.2
  Rename PixelColour to PixelColor for consistency. - All other instances use American spelling of color.
  Update copyright notice.
  List changes in change log.
  Update copyright notice.
  Remove redundant virtual where we have virtual and override specified.
  ...

# Conflicts:
#	LICENSE.md
#	source/gwork/include/Gwork/Controls/Base.h
#	source/gwork/include/Gwork/Controls/CheckBox.h
#	source/gwork/include/Gwork/Controls/ComboBox.h
#	source/gwork/include/Gwork/Controls/HorizontalScrollBar.h
#	source/gwork/include/Gwork/Controls/MenuItem.h
#	source/gwork/include/Gwork/Controls/ResizableControl.h
#	source/gwork/include/Gwork/Controls/ScrollBar.h
#	source/gwork/include/Gwork/Controls/ScrollControl.h
#	source/gwork/include/Gwork/Controls/TextBox.h
#	source/gwork/include/Gwork/Controls/VerticalScrollBar.h
#	source/gwork/include/Gwork/Skins/Simple.h
#	source/gwork/include/Gwork/Skins/TexturedBase.h
#	source/gwork/source/Events.cpp
#	source/test/source/api/Slider.cpp
#	source/util/include/Gwork/Util/ControlReflect.h
#	source/util/source/ControlFactory/FilePicker_Factory.cpp
#	source/util/source/ControlFactory/Image_Factory.cpp
#	source/util/source/ControlFactory/Label_Factory.cpp
#	source/util/source/ControlFactory/ListBox_Factory.cpp
#	source/util/source/ControlFactory/MenuStrip_Factory.cpp
#	source/util/source/ControlFactory/PageControl_Factory.cpp
#	source/util/source/ControlFactory/ProgressBar_Factory.cpp
#	source/util/source/ControlFactory/Rectangle_Factory.cpp
#	source/util/source/ControlFactory/Slider_Factory.cpp
#	source/util/source/ControlFactory/TextBox_Factory.cpp
#	source/util/source/ControlFactory/VerticalSplitter_Factory.cpp

+++ source/gwork/source/BaseRender.cpp
@@ -1,7 +1,7 @@
 /*
  *  Gwork
  *  Copyright (c) 2010 Facepunch Studios
- *  Copyright (c) 2013-16 Billy Quith
+ *  Copyright (c) 2013-17 Nick Trout
  *  See license in Gwork.h
  */
 
@@ -190,7 +190,7 @@ void Base::RenderText(Gwk::Font* font, Gwk::Point pos, const Gwk::String& text)
 Gwk::Point Base::MeasureText(Gwk::Font* font, const Gwk::String& text)
 {
     Gwk::Point p;
-    p.x = font->size*Scale()*(float)text.length()*0.4f;
+    p.x = font->size*Scale() * text.length() * 0.4f;
     p.y = font->size*Scale();
     return p;
 }
